requirements.txt
Django==1.7.4
Django-mptt==0.6
psycopg2
django-allauth
whitenoise
